+++ test/unit/standalone-wallets.ts
@@ -6,6 +6,7 @@ import { utils } from "ethers"
 import { WalletUtil } from "../../src/util/signers"
 import { Random } from "../../src/util/random"
 import { JsonSignature, BytesSignature } from "../../src/util/data-signing"
+import { compressPublicKey } from "../../dist"
 
 addCompletionHooks()
 
@@ -45,7 +46,8 @@ describe("Standalone Ethereum wallets", () => {
 
         const wallet = WalletUtil.fromSeededPassphrase(passphrase, hexSeed)
         expect(wallet.privateKey).to.eq("0x58c6192cbffe39d20f6dbaa2957a6d6a4116489a2bb66caab5c4a0bfa83d887b")
-        expect(wallet["_signingKey"]().publicKey).to.eq("0x04de6d532b6979899729f9e98869888ea7fdbc446f9f3ea732d23c7bcd10c784d041887d48ebc392c4ff51882ae569ca1553f6ab6538664bced6cca6855acbbade")
+        expect(compressPublicKey(wallet.publicKey)).to.eq("0x02de6d532b6979899729f9e98869888ea7fdbc446f9f3ea732d23c7bcd10c784d0")
+        expect(wallet.publicKey).to.eq("0x04de6d532b6979899729f9e98869888ea7fdbc446f9f3ea732d23c7bcd10c784d041887d48ebc392c4ff51882ae569ca1553f6ab6538664bced6cca6855acbbade")
         expect(await wallet.getAddress()).to.eq("0xf76564CBF51B1F050c84fC01400088ACD2704F2e")
 
         const msg = utils.toUtf8Bytes("Hello")
@@ -137,6 +139,7 @@ describe("Standalone Ethereum wallets", () => {
 
         const signature = await JsonSignature.sign(jsonBody, wallet)
 
-        expect(JsonSignature.isValid(signature, wallet["_signingKey"]().publicKey, jsonBody)).to.be.true
+        expect(JsonSignature.isValid(signature, compressPublicKey(wallet.publicKey), jsonBody)).to.be.true
+        expect(JsonSignature.isValid(signature, wallet.publicKey, jsonBody)).to.be.true
     })
 })